chore(adv): checkpoint tk-afbc111de8e8
Change: fixWorkflowReliabilityDefects
Task: tk-afbc111de8e8
Mode: complete
Verification: Harden remediation: unified MutationApplicationUnconfirmedError in shared utility, removed dead coverage helper/test import, clarified root-session wording. tr_mrsklr1h_1a5c2417 passed 183 targeted tests; tr_mrsklgez_802e9f76 passed full project check. Final full suite passed 422 files / 6501 tests + 1 expected fail; build tr_mrsjwj5x_2052cd81 passed. Unauthorized direct spec write was reverted because archive remains sole global-spec writer.
